Fix testing bugs and refactor for better code management for Service Profiles
